.history/src/requirements_20211115000743.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
djangorestframework==3.12.4
drf-jwt==1.19.1
gunicorn==20.1.0
idna==3.3
Pillow==8.3.2
PyJWT==2.2.0
python-decouple==3.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
twilio==7.1.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
rsa
cryptography